Convert the repo's ownership and add maven.
Since this repo will be a collection of updates and integrations to the
randomizer, it should be clear that the ownership has changed. This
changes the version to 2.0.0 and changes all of the dabomstew parts that
I have found so far. There will likely need to be a few more updates of
this ilk. Also Maven is added so that it builds with Java 11.